+++ .changeset/metadata-panel-styling.md
@@ -2,4 +2,4 @@
 '@workflow/web-shared': patch
 ---
 
-Align the detail panel top info rows with the Attributes section styling, group them in a collapsible Metadata section, and render values in monospace. Extend `cn` with custom tailwind-merge groups for design-system typography utilities.
+Align the detail panel top info rows with the Attributes section styling, group them in a collapsible Metadata section, and render values in monospace.
