+++ packages/web-shared/src/components/new-trace-viewer/components/copy-button.tsx
@@ -2,7 +2,7 @@
 
 import { Check, Copy } from 'lucide-react';
 import { type JSX, useEffect, useRef, useState } from 'react';
-import { cn } from '../../../lib/cn';
+import { cn } from '../../../lib/utils';
 
 interface CopyButtonProps {
   copyText: string;
