+++ packages/web-shared/src/components/new-trace-viewer/components/event-list.tsx
@@ -1,6 +1,6 @@
 import { Circle } from 'lucide-react';
 import { useRef } from 'react';
-import { cn } from '../../../lib/cn';
+import { cn } from '../../../lib/utils';
 import type { Span } from '../types';
 import { formatDurationPrecise } from '../../trace-viewer/util/timing';
 import {
