+++ packages/web-shared/src/components/new-trace-viewer/components/middle-truncate/middle-truncate.tsx
@@ -2,7 +2,7 @@
 
 import { useCallback, useRef } from 'react';
 import type { ClipboardEvent, ComponentPropsWithoutRef, JSX } from 'react';
-import { cn } from '../../../../lib/cn';
+import { cn } from '../../../../lib/utils';
 import {
   getMiddleTruncateCopyText,
   getMiddleTruncateCopyTextFromSelectionText,
