+++ packages/web-shared/src/components/new-trace-viewer/components/split-pane.tsx
@@ -10,7 +10,7 @@ import {
   useRef,
   useState,
 } from 'react';
-import { cn } from '../../../lib/cn';
+import { cn } from '../../../lib/utils';
 
 const GUTTER_PX = 1;
 const MIN_PX = 50;
