+++ packages/web-shared/src/components/new-trace-viewer/components/timeline.tsx
@@ -11,7 +11,7 @@ import {
   useRef,
   useState,
 } from 'react';
-import { cn } from '../../../lib/cn';
+import { cn } from '../../../lib/utils';
 import {
   formatDuration,
   formatDurationPrecise,
