+++ packages/web-shared/src/components/ui/button.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { cva, type VariantProps } from 'class-variance-authority';
-import { cn } from '../../lib/cn';
+import { cn } from '../../lib/utils';
 
 const buttonVariants = cva(
   [
