+++ packages/web-shared/src/components/ui/card.tsx
@@ -2,7 +2,7 @@
 
 import * as React from 'react';
 
-import { cn } from '../../lib/cn';
+import { cn } from '../../lib/utils';
 
 const Card = React.forwardRef<
   HTMLDivElement,
