+++ packages/web-shared/src/components/ui/collapsible.tsx
@@ -8,7 +8,7 @@ import {
   useContext,
   useState,
 } from 'react';
-import { cn } from '../../lib/cn';
+import { cn } from '../../lib/utils';
 
 type CollapsibleVariant = 'section' | 'card';
