+++ packages/web-shared/src/components/ui/error-card.tsx
@@ -2,7 +2,7 @@
 
 import { AlertCircle, ChevronDown } from 'lucide-react';
 import { useState } from 'react';
-import { cn } from '../../lib/cn';
+import { cn } from '../../lib/utils';
 
 interface ErrorCardProps {
   /** Title shown in the header */
