+++ packages/web-shared/src/components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../../lib/cn';
+import { cn } from '../../lib/utils';
 
 function Skeleton({
   className,
