+++ packages/web-shared/src/components/ui/tooltip.tsx
@@ -2,7 +2,7 @@
 
 import * as TooltipPrimitive from '@radix-ui/react-tooltip';
 import * as React from 'react';
-import { cn } from '../../lib/cn';
+import { cn } from '../../lib/utils';
 
 const TooltipProvider = TooltipPrimitive.Provider;
