+++ packages/web-shared/src/lib/utils.ts
@@ -1,5 +1,11 @@
 import type { Step } from '@workflow/world';
 import type { ModelMessage } from 'ai';
+import { type ClassValue, clsx } from 'clsx';
+import { twMerge } from 'tailwind-merge';
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
 
 const MS_IN_SECOND = 1000;
 const MS_IN_MINUTE = 60 * MS_IN_SECOND;
